Started implementation of business logic on the repo/service layers

+++ src/services/EventService.ts
@@ -24,14 +24,25 @@ export class EventService {
         return await this.eventRepo.getEventsHostedByUserById(userId);
     }
 
+    public async exists(id: number) {
+        return await this.eventRepo.exists(id);
+    }
+
     public async createEvent(
         hostId: number,
         dateHosted: Date
     ) {
+        // Not sure if we need an exists check here, as the same host could host multiple events in one day
         return await this.eventRepo.create(hostId, dateHosted);
     }
 
     public async deleteEvent(id: number) {
+        const exists = await this.exists(id);
+
+        // During the implementation of the error class story, this should throw an error
+        // with an appropriate status code, where this error would be handled in the controller
+        if (!exists) return true;
+
         return await this.eventRepo.delete(id);
     }
 
@@ -39,6 +50,10 @@ export class EventService {
         eventId: number,
         update: IEventUpdateDto
     ) {
+        const exists = await this.exists(eventId);
+
+        // This should throw an error, which should be implemented during the error class implementation story
+        if (!exists) return false;
         return await this.eventRepo.update(eventId, update);
     }
 }
